Added functions to clear data in ScheduleModal
